Implement podcast permission checking

+++ routes/podcast/delete.py
@@ -1,5 +1,6 @@
 from flask import redirect, session, render_template, abort
 
+from models.user import User
 from routes.podcast import podcast_blueprint
 
 from models.podcast import Podcast
@@ -9,13 +10,18 @@ from util import user_logged_in
 
 @podcast_blueprint.route("/<int:podcast_id>/delete")
 def delete(podcast_id):
-    # todo: flashing
     if not user_logged_in():
         abort(401)
 
-    # todo: validate owner or admin, err 403
+    user = User.get_from_id(session['user_id'])
+
     try:
         podcast = Podcast.get_from_id(podcast_id)
+        has_perm = (user is not None) and (
+            user.is_admin or (user.id_ == podcast.maintainer)
+        )
+        if not has_perm:
+            abort(403)
         podcast.delete()
     except:
         abort(500)
